Create sudokuSolver2.cpp
This algorithm is also not perfect. It does a little more than the first; it checks if there is no other possible place for a number within a given row, col, or groupCell. It still can't do harder puzzles, but can handle some medium puzzles that the first struggled with.